changed from parseint to parsefloat

+++ calculator.js
@@ -17,27 +17,27 @@ const multiplyBtn = document.getElementById("multiply");
 const divideBtn = document.getElementById("divide");
 
 addBtn.addEventListener("click", function() {
-    const number1 = parseInt(document.getElementById("number1").value);
-    const number2 = parseInt(document.getElementById("number2").value);
+    const number1 = parseFloat(document.getElementById("number1").value);
+    const number2 = parseFloat(document.getElementById("number2").value);
     const result = add(number1, number2);
     document.getElementById("calculation-result").textContent = result;
 });
 
 subtractBtn.addEventListener("click", function() {
-    const number1 = parseInt(document.getElementById("number1").value);
-    const number2 = parseInt(document.getElementById("number2").value);
+    const number1 = parseFloat(document.getElementById("number1").value);
+    const number2 = parseFloat(document.getElementById("number2").value);
     const result = subtract(number1, number2);
     document.getElementById("calculation-result").textContent = result;
 });
 multiplyBtn.addEventListener("click", function() {
-    const number1 = parseInt(document.getElementById("number1").value);
-    const number2 = parseInt(document.getElementById("number2").value);
+    const number1 = parseFloat(document.getElementById("number1").value);
+    const number2 = parseFloat(document.getElementById("number2").value);
     const result = multiply(number1, number2);
     document.getElementById("calculation-result").textContent = result;
 });
 divideBtn.addEventListener("click", function() {
-    const number1 = parseInt(document.getElementById("number1").value);
-    const number2 = parseInt(document.getElementById("number2").value);
+    const number1 = parseFloat(document.getElementById("number1").value);
+    const number2 = parseFloat(document.getElementById("number2").value);
     if (number2 === 0) {
         document.getElementById("calculation-result").textContent = "Cannot divide by zero";
         return;
